# store: avoid copying same info twice with caching bucket