Currently we always do the same regardless of whether caching bucket is
enabled or not. It leads to some inefficiencies because memory returned
by caching bucket is already resident in the process and it's not
temporary i.e. not retrieved from a sync.Pool, so it means that we can
skip over the "saving" part and use that memory directly.

Extra allocs come from the part where `readOffset` + `n` leads to a
situation where we need to concatenate two slices in `readNumBytes`.

Overall, given the high variability of allocs benchmarks, I think this
is definitely an improvement. A surprising high amount of CPU time is
spent just copying bytes.

Diff:

```
name                                         old time/op    new time/op    delta
BlockSeriesCachingBucket/concurrency:_1-16     12.7ms ± 8%    12.1ms ±10%   -5.06%  (p=0.000 n=27+30)
BlockSeriesCachingBucket/concurrency:_2-16     8.42ms ± 7%    7.64ms ± 5%   -9.23%  (p=0.000 n=28+29)
BlockSeriesCachingBucket/concurrency:_4-16     6.66ms ±10%    5.43ms ± 8%  -18.54%  (p=0.000 n=30+27)
BlockSeriesCachingBucket/concurrency:_8-16     6.39ms ±10%    5.68ms ±16%  -11.09%  (p=0.000 n=29+30)
BlockSeriesCachingBucket/concurrency:_16-16    11.4ms ±15%     8.4ms ±14%  -26.98%  (p=0.000 n=30+30)
BlockSeriesCachingBucket/concurrency:_32-16    12.2ms ±14%     7.9ms ±16%  -35.22%  (p=0.000 n=30+29)

name                                         old alloc/op   new alloc/op   delta
BlockSeriesCachingBucket/concurrency:_1-16     36.6MB ± 4%    31.1MB ± 5%  -14.99%  (p=0.000 n=30+30)
BlockSeriesCachingBucket/concurrency:_2-16     36.4MB ± 7%    31.8MB ± 3%  -12.60%  (p=0.000 n=30+29)
BlockSeriesCachingBucket/concurrency:_4-16     34.5MB ± 9%    33.1MB ± 3%   -4.00%  (p=0.001 n=30+30)
BlockSeriesCachingBucket/concurrency:_8-16     34.5MB ±10%    35.3MB ± 5%     ~     (p=0.084 n=29+29)
BlockSeriesCachingBucket/concurrency:_16-16    56.1MB ± 4%    49.8MB ± 6%  -11.30%  (p=0.000 n=29+30)
BlockSeriesCachingBucket/concurrency:_32-16    59.6MB ± 7%    49.8MB ± 9%  -16.47%  (p=0.000 n=30+30)

name                                         old allocs/op  new allocs/op  delta
BlockSeriesCachingBucket/concurrency:_1-16      52.9k ± 4%     53.0k ± 5%     ~     (p=0.865 n=30+30)
BlockSeriesCachingBucket/concurrency:_2-16      54.0k ± 3%     54.1k ± 3%     ~     (p=0.905 n=28+29)
BlockSeriesCachingBucket/concurrency:_4-16      54.1k ± 8%     56.4k ± 4%   +4.21%  (p=0.001 n=30+30)
BlockSeriesCachingBucket/concurrency:_8-16      55.4k ±12%     60.6k ± 4%   +9.28%  (p=0.000 n=29+29)
BlockSeriesCachingBucket/concurrency:_16-16     92.5k ± 0%     86.9k ± 6%   -6.02%  (p=0.000 n=29+30)
BlockSeriesCachingBucket/concurrency:_32-16      103k ± 0%       93k ± 8%   -9.81%  (p=0.000 n=30+30)
```


## Verification

`make test-local` passes